Reorganize code and try to sort out forward references
