+++ nanomesh/mesh2d/__init__.py
@@ -1,7 +1,9 @@
-from .helpers import pad, simple_triangulate
-from .mesher import Mesher2D, generate_2d_mesh
+from nanomesh.mesh2d.helpers import (compare_mesh_with_image, pad,
+                                     simple_triangulate)
+from nanomesh.mesh2d.mesher import Mesher2D, generate_2d_mesh
 
 __all__ = [
+    'compare_mesh_with_image',
     'generate_2d_mesh',
     'Mesher2D',
     'pad',
