+++ nanomesh/plane.py
@@ -5,9 +5,9 @@ from typing import Callable, Union
 import matplotlib.pyplot as plt
 import numpy as np
 
-from .base_image import BaseImage
-from .mesh_container import TriangleMesh
-from .utils import show_image
+from nanomesh.base_image import BaseImage
+from nanomesh.mesh_container import TriangleMesh
+from nanomesh.utils import show_image
 
 logger = logging.getLogger(__name__)
 
@@ -86,7 +86,7 @@ class Plane(BaseImage):
         mesh : TriangleMesh
             Description of the mesh.
         """
-        from .mesh2d import generate_2d_mesh
+        from nanomesh.mesh2d import generate_2d_mesh
         return generate_2d_mesh(image=self.image, **kwargs)
 
     def select_roi(self, from_points: np.ndarray = None):
@@ -97,7 +97,7 @@ class Plane(BaseImage):
         roi : `nanomesh.roi2d.ROISelector`
             Region of interest object. Bounding box is stored in `roi.bbox`.
         """
-        from .roi2d import ROISelector
+        from nanomesh.roi2d import ROISelector
         ax = self.show(title='Select region of interest')
         if from_points is not None:
             # reverse columns to match image
@@ -135,7 +135,7 @@ class Plane(BaseImage):
         `nanomesh.Plane`
             Cropped region as `nanomesh.Plane` object.
         """
-        from .roi2d import extract_rectangle
+        from nanomesh.roi2d import extract_rectangle
         cropped = extract_rectangle(self.image, bbox=bbox)
         return Plane(cropped)
 
@@ -151,7 +151,7 @@ class Plane(BaseImage):
         -------
         ax : matplotlib.Axes
         """
-        from .mesh_utils import compare_mesh_with_image
+        from nanomesh.mesh_utils import compare_mesh_with_image
         return compare_mesh_with_image(image=self.image, mesh=mesh)
 
     def compare_with_digitized(self, digitized: Union[np.ndarray, 'Plane'],
