+++ tests/test_mesh_utils.py
@@ -1,8 +1,8 @@
 import numpy as np
 from matplotlib.testing.decorators import image_comparison
 
+from nanomesh.mesh2d import compare_mesh_with_image
 from nanomesh.mesh_container import TriangleMesh
-from nanomesh.mesh_utils import compare_mesh_with_image
 
 
 @image_comparison(
